# Ensure numba tokenization does not use slow pickle path