--- a/dask/tests/test_tokenize.py
+++ b/dask/tests/test_tokenize.py
@@ -1369,6 +1369,12 @@ all_numba_funcs = [
 def test_tokenize_numba(func):
     assert func(1, 2) == 3
     check_tokenize(func)
+    for func in all_numba_funcs:
+        tokens = normalize_token(func)
+
+        # Ensure that we attempt to tokenize it instead of dumping it into pickle
+        assert isinstance(tokens, tuple)
+        assert isinstance(tokens[1], tuple)
 
 
 @pytest.mark.skipif("not numba")