Closes https://github.com/dask/dask/issues/11416

The example provided in the issue

```python
%%time
for i in range(3):
    a = da.ones((10000, 1000, 10), chunks=(1000, 1000, 10), dtype=np.int8)
    res = da.map_blocks(inc, a, dtype=np.int8).compute()
```

now runs on my machine in ~150ms instead of 2.6s

Isolating this to the tokenization of the function

```python
%timeit tokenize(inc)

# main - 345 ms ± 9.06 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# PR - 447 μs ± 4.36 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

```

so tokenization is about three orders of magnitude faster